requirements.txt
amqp==2.1.1
backports.shutil-get-terminal-size==1.0.0
billiard==3.5.0.2
blinker==1.4
celery==4.0.0
construct==2.8.8
decorator==4.0.10
Django==1.10.3
django-braces==1.10.0
django-redis==4.6.0
djangorestframework==3.5.3
enum34==1.1.6
future==0.16.0
ipython==5.1.0
ipython-genutils==0.1.0
kombu==4.0.0
pathlib2==2.1.0
pefile==2016.3.28
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.9
psycopg2==2.6.2
ptyprocess==0.5.1
Pygments==2.1.3
pysolr==3.6.0
python-ptrace==0.9.1
pytz==2016.7
redis==2.10.5
requests==2.12.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
vine==1.1.3
wcwidth==0.1.7
wheel==0.24.0
